Fixed XML problems and added sections to the ExampleRobotHardware class
So what happened with the XML problems in the previous release - where the phone was unable to locate the resources - was that it tried to grab resources statically which won't work when compiled on the computer and used on the phone. However, enums are still static, so what I did is split it up - you pass the R.string.whatever id to the enumerator and then the init() function calls functions that pass each instance the hardware map and actually fetches the resources. All of the getResource[String,Int,Whatever] are now non-static again, because they have to use HardwareMap.getResources().